+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # meseret change log
 
 ## v1.2.0
+- Construct Koa app when `ServerApp` is constructed, before calling `ServerApp.start()`.
 - Add key and session support.
 - Improved README.md and package.json.